Multiple critical bugs in 3-opt search
I think I found several critical bugs in the ThreeOpt code. Could you please verify that I'm actually on to something?

Please assume that the route nodes are indexed n0-n1-n2-... the original implementation started from edges n0-n1 n2-n3 n4-n5, which caused the edges n1-n2 n3-n4 etc. not to be checked.  The same bug was also in the O(n^3) main search loop, where the edges between consecutive nodes were similarly skipped due to how.

Also, the end nodes for the loop were incorrect (too many lookups from the prev_array) as the search was terminated before reaching the final nodes of the route. As an result 3-opt did not work as expected and produced wrong (or severely wrong, depending on the input) results.

There is an additional issue, which IS NOT CURRENTLY FIXED: the route[r].end and route[r].start are not (!) updated by the ThreeOpt::move. Therefore, it is possible that a move is accepted and the subsequent 3-opt searches are incolplete. A workaround is to call VRP::normalize_route_numbers() after each successful search.